api/requirements.txt
fastapi==0.110.1
uvicorn==0.18.3
python-multipart==0.0.5
python-magic==0.4.27
celery[redis]==5.3.6
pydantic-settings>=2.0.0
